requirements.txt
discord.py>=2.3.0
midtransclient>=1.4.0
Flask>=2.3.0
pytz
openpyxl>=3.0.0
requests
psycopg2-binary>=2.9.0
